Create bower_components and node_modules if needed on build
The WriteLinesToFile task used in bower.targets and node.targets fails
if the bower_components or node_modules folders themselves don't exist.
If any npm or bower packages have already been installed, it's not an
issue, but on an initial install of these NuGet packages (before npm or
bower are used), the solution won't build properly.